# openapi/swagger.yaml
openapi: 3.0.1
info:
  title: Process management
  description: Process management
  version: 1.0.0
servers:
- url: /
paths:
...
  /process/trips/{id}/fix:
    post:
      tags:
        - Process
      operationId: fixTrip
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
...
  /mountStatus/{id}:
    get:
      tags:
        - MountStatus
      operationId: getMountStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountStatus'
    post:
      tags:
        - MountStatus
      operationId: createMountStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MountStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountStatus'
    put:
      tags:
        - MountStatus
      operationId: updateMountStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MountStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountStatus'
    delete:
      tags:
        - MountStatus
      operationId: deleteMountStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
...
  /chargingStationStatus/{id}:
    get:
      tags:
        - ChargingStationStatus
      operationId: getChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    post:
      tags:
        - ChargingStationStatus
      operationId: createChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    put:
      tags:
        - ChargingStationStatus
      operationId: updateChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    delete:
      tags:
        - ChargingStationStatus
      operationId: deleteChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
...
    MountStatus:
      required:
        - id
        - state
        - timestamp
      type: object
      properties:
        id:
          type: string
        mountState:
          type: string
          enum:
            - Free
            - InUse
            - Disrupted
        boxId:
          type: string
        timestamp:
          type: string
          format: date-time
    ChargingStationStatus:
      required:
        - id
        - state
        - timestamp
      type: object
      properties:
        id:
          type: string
        chargingStationState:
          type: string
          enum:
            - Free
            - InUse
            - Disrupted
        vehicleId:
          type: string
        timestamp:
          type: string
          format: date-time